move content to body and optimize memorize

+++ Entities/Quiz.cs
@@ -4,7 +4,7 @@
     {
         public required string Question { get; set; }
         public required List<string> Options { get; set; }
-        public required short RightOptionIndex { get; set; }
+        public required sbyte RightOptionIndex { get; set; }
         public required string ExplanationInVietnamese { get; set; }
     }
 }
